requirements.txt
langchain
python-dotenv
langchain-community
pypdf
beautifulsoup4
arxiv
pymupdf
langchain-text-splitters
langchain-openai
langchain-ollama
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
streamlit
langchain-groq
fastapi
uvicorn
langserve[all]
wikipedia